2 nody prepinane definem
